Part 2- refactpring boiler plate
