+++ backend/de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/apps/EMailDialog.java
@@ -30,8 +30,8 @@ import de.metas.email.mailboxes.UserEMailConfig;
 import de.metas.i18n.IMsgBL;
 import de.metas.i18n.ITranslatableString;
 import de.metas.letters.model.I_AD_BoilerPlate;
-import de.metas.letters.model.MADBoilerPlate;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.logging.LogManager;
 import de.metas.user.api.IUserBL;
 import de.metas.user.api.IUserDAO;
