+++ backend/de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/grid/ed/VTextLong.java
@@ -29,7 +29,7 @@ import org.compiere.grid.ed.menu.EditorContextPopupMenu;
 import org.compiere.model.GridField;
 import org.compiere.swing.CTextPane;
 
-import de.metas.letters.model.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate;
 
 /**
  *	Long Text (CBLOB) Editor
