+++ backend/de.metas.business/src/main/java/de/metas/boilerplate/BoilerPlate.java
@@ -1,18 +1,8 @@
-package de.metas.letter;
-
-import de.metas.i18n.Language;
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
-import org.adempiere.ad.expression.api.IExpressionEvaluator;
-import org.adempiere.ad.expression.api.IStringExpression;
-import org.compiere.util.Evaluatee;
-
 /*
  * #%L
- * de.metas.adempiere.adempiere.base
+ * de.metas.business
  * %%
- * Copyright (C) 2018 metas GmbH
+ * Copyright (C) 2023 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -21,14 +11,24 @@ import org.compiere.util.Evaluatee;
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
+
+package de.metas.boilerplate;
+
+import de.metas.i18n.Language;
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
+import org.adempiere.ad.expression.api.IExpressionEvaluator;
+import org.adempiere.ad.expression.api.IStringExpression;
+import org.compiere.util.Evaluatee;
 
 /**
  * @author metas-dev <dev@metasfresh.com>
