+++ backend/de.metas.business/src/main/java/de/metas/boilerplate/MADBoilerPlateRef.java
@@ -1,34 +1,34 @@
-/**
- * 
- */
-package de.metas.letters.model;
-
 /*
  * #%L
- * de.metas.adempiere.adempiere.base
+ * de.metas.business
  * %%
- * Copyright (C) 2015 metas GmbH
+ * Copyright (C) 2023 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
+/**
+ * 
+ */
+package de.metas.boilerplate;
 
 import java.sql.ResultSet;
 import java.util.Properties;
 
+import de.metas.letters.model.X_AD_BoilerPlate_Ref;
 import org.adempiere.exceptions.AdempiereException;
 
 /**
