+++ backend/de.metas.business/src/main/java/de/metas/boilerplate/POSourceDocument.java
@@ -1,0 +1,46 @@
+/*
+ * #%L
+ * de.metas.business
+ * %%
+ * Copyright (C) 2023 metas GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation, either version 2 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program. If not, see
+ * <http://www.gnu.org/licenses/gpl-2.0.html>.
+ * #L%
+ */
+
+package de.metas.boilerplate;
+
+import lombok.AllArgsConstructor;
+import lombok.NonNull;
+import org.compiere.model.PO;
+
+@AllArgsConstructor
+private static final class POSourceDocument implements SourceDocument
+{
+	@NonNull
+	private final PO po;
+
+	@Override
+	public boolean hasFieldValue(final String fieldName)
+	{
+		return po.get_ColumnIndex(fieldName) >= 0;
+	}
+
+	@Override
+	public Object getFieldValue(final String fieldName)
+	{
+		return po.get_Value(fieldName);
+	}
+}
