+++ backend/de.metas.business/src/main/java/de/metas/order/model/interceptor/C_Order.java
@@ -37,9 +37,9 @@ import de.metas.i18n.ITranslatableString;
 import de.metas.i18n.Language;
 import de.metas.interfaces.I_C_OrderLine;
 import de.metas.invoice.service.IInvoiceBL;
-import de.metas.letter.BoilerPlate;
-import de.metas.letter.BoilerPlateId;
-import de.metas.letter.BoilerPlateRepository;
+import de.metas.boilerplate.BoilerPlate;
+import de.metas.boilerplate.BoilerPlateId;
+import de.metas.boilerplate.BoilerPlateRepository;
 import de.metas.order.DeliveryViaRule;
 import de.metas.order.IOrderBL;
 import de.metas.order.IOrderDAO;
