+++ backend/de.metas.business/src/main/java/de/metas/shipping/model/validator/ShipperTransportationMailNotification.java
@@ -33,8 +33,8 @@ import de.metas.email.EMailAddress;
 import de.metas.email.EMailSentStatus;
 import de.metas.email.impl.EMailSendException;
 import de.metas.letters.model.IEMailEditor;
-import de.metas.letters.model.MADBoilerPlate;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.logging.LogManager;
 import de.metas.shipping.model.I_M_ShippingPackage;
 import de.metas.shipping.model.MMShipperTransportation;
