+++ backend/de.metas.document.archive/de.metas.document.archive.base/src/main/java/de/metas/document/archive/async/spi/impl/MailWorkpackageProcessor.java
@@ -25,9 +25,9 @@ import de.metas.email.mailboxes.UserEMailConfig;
 import de.metas.i18n.AdMessageKey;
 import de.metas.i18n.IMsgBL;
 import de.metas.i18n.Language;
-import de.metas.letter.BoilerPlate;
-import de.metas.letter.BoilerPlateId;
-import de.metas.letter.BoilerPlateRepository;
+import de.metas.boilerplate.BoilerPlate;
+import de.metas.boilerplate.BoilerPlateId;
+import de.metas.boilerplate.BoilerPlateRepository;
 import de.metas.organization.OrgId;
 import de.metas.process.AdProcessId;
 import de.metas.process.ProcessExecutor;
