+++ backend/de.metas.marketing/serialletter/src/main/java/de/metas/letter/service/async/spi/impl/C_Letter_CreateFromMKTG_ContactPerson_Async.java
@@ -8,8 +8,8 @@ import de.metas.bpartner.BPartnerLocationId;
 import de.metas.document.location.DocumentLocation;
 import de.metas.document.location.RenderedAddressAndCapturedLocation;
 import de.metas.document.location.impl.DocumentLocationBL;
-import de.metas.letter.BoilerPlate;
-import de.metas.letter.BoilerPlateRepository;
+import de.metas.boilerplate.BoilerPlate;
+import de.metas.boilerplate.BoilerPlateRepository;
 import de.metas.letters.model.Letter;
 import de.metas.letters.model.LetterRepository;
 import de.metas.marketing.base.model.ContactPerson;
