+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/letters/api/impl/TextTemplateBL.java
@@ -4,15 +4,15 @@ import de.metas.bpartner.service.IBPartnerBL;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.cache.annotation.CacheCtx;
 import de.metas.document.location.IDocumentLocationBL;
-import de.metas.letter.BoilerPlateId;
+import de.metas.boilerplate.BoilerPlateId;
 import de.metas.letters.api.ITextTemplateBL;
 import de.metas.letters.model.I_AD_BoilerPlate;
 import de.metas.letters.model.I_C_Letter;
 import de.metas.letters.model.I_T_Letter_Spool;
 import de.metas.letters.model.Letter;
 import de.metas.letters.model.LetterDocumentLocationAdapter;
-import de.metas.letters.model.MADBoilerPlate;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.letters.spi.ILetterProducer;
 import de.metas.logging.LogManager;
 import de.metas.organization.OrgId;
