+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/letters/model/LettersValidator.java
@@ -22,11 +22,14 @@ package de.metas.letters.model;
  * #L%
  */
 
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlateVar;
+import de.metas.boilerplate.MADBoilerPlateVarEval;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.interfaces.I_M_Movement;
 import de.metas.letters.api.ITextTemplateBL;
 import de.metas.letters.api.impl.TextTemplateBL;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import org.adempiere.model.InterfaceWrapperHelper;
