+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/letters/report/AD_BoilderPlate_SendToGroup.java
@@ -30,8 +30,8 @@ import de.metas.email.EMail;
 import de.metas.email.EMailAddress;
 import de.metas.email.EMailSentStatus;
 import de.metas.letters.model.IEMailEditor;
-import de.metas.letters.model.MADBoilerPlate;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.logging.LogManager;
 import de.metas.process.JavaProcess;
 import de.metas.process.ProcessInfoParameter;
