+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/org/compiere/report/email/service/impl/EmptyEmailParams.java
@@ -3,7 +3,7 @@ package org.compiere.report.email.service.impl;
 import org.compiere.model.I_AD_User;
 import org.compiere.report.email.service.IEmailParameters;
 
-import de.metas.letters.model.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate;
 
 /**
  * Default implementation. Always returns <code>null</code>.
