+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/letter/LetterRestController.java
@@ -8,12 +8,12 @@ import de.metas.document.location.DocumentLocation;
 import de.metas.document.location.IDocumentLocationBL;
 import de.metas.document.location.RenderedAddressAndCapturedLocation;
 import de.metas.i18n.IMsgBL;
-import de.metas.letter.BoilerPlateId;
+import de.metas.boilerplate.BoilerPlateId;
 import de.metas.letters.api.ITextTemplateBL;
 import de.metas.letters.model.Letter;
 import de.metas.letters.model.Letters;
-import de.metas.letters.model.MADBoilerPlate;
-import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
+import de.metas.boilerplate.MADBoilerPlate;
+import de.metas.boilerplate.MADBoilerPlate.BoilerPlateContext;
 import de.metas.ui.web.config.WebConfig;
 import de.metas.ui.web.letter.WebuiLetter.WebuiLetterBuilder;
 import de.metas.ui.web.letter.json.JSONLetter;
