+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/session/InternalUserSessionData.java
@@ -3,7 +3,7 @@ package de.metas.ui.web.session;
 import com.google.common.base.MoreObjects;
 import de.metas.contracts.ConditionsId;
 import de.metas.i18n.Language;
-import de.metas.letter.BoilerPlateId;
+import de.metas.boilerplate.BoilerPlateId;
 import de.metas.organization.OrgId;
 import de.metas.security.RoleId;
 import de.metas.ui.web.session.json.WebuiSessionId;
